requirements.txt
pybitcoin
